refactor filter chain match options
